requirements.txt
selenium==4.16.0
webdriver_manager==4.0.1
python-dotenv==1.0.0
pandas==2.1.4
beautifulsoup4==4.12.2
requests==2.31.0
browser-cookie3==0.19.1
PyJWT==2.8.0
textual==0.44.1
rich==13.7.0
typer[all]==0.9.0
aiohttp==3.10.0
asyncio==3.4.3
